requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --generate-hashes requirements.in
#

pandas
numpy
SQLAlchemy
psycopg2-binary
python-dotenv
yfinance
requests
pyarrow
#numpy==1.26.4
#pandas==2.1.4
#SQLAlchemy==1.4.52
#psycopg2-binary==2.9.9
#python-dotenv==1.0.1
#yfinance==0.2.65
#requests==2.31.0
#pyarrow==16.1.0
pytest==7.4.4
pytest-cov==5.0.0
streamlit>=1.28.0
plotly>=5.17.0

